+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes for the solr-cmd-utils project will be documented in this fi
 - implement a document reader
 - implement a document writer
 - add a mapping function for base65 encoding and decoding
+- make it possible to read multiple csv files
+- implement a SearchQueryAnalyzerFilter
+- implement a document joiner filter
+- implement a html reader for testing
 
 ### Changed