+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes for the solr-cmd-utils project will be documented in this fi
 
 ## [Unreleased]
 ### Added
+
+### Changed
+
+
+### Fixed
+
+
+### Deprecated
+
+
+## [3.17]
+### Added
 - make it possible to log documents fields in the status filter
 - implement a document reader
 - implement a document writer
@@ -20,6 +32,8 @@ All notable changes for the solr-cmd-utils project will be documented in this fi
 
 
 ### Deprecated
+
+
 ## [3.16] - 2021-02-20
 ### Added
 - add a cache max age parameter to the http filter